+++ assets/src/js/999-scripts.js
@@ -24,5 +24,49 @@ jQuery(function($){
 			// }, 700);
 		// });
 	// }
+	
+	// Shortcode gallery
+	$('[data-fancybox="gallery"]').fancybox({
+		loop: true,
+		buttons: ['close']
+	});
+
+	// Scroll to accordion top after the panel is opened
+	$('.nx-accordion').on('shown.bs.collapse', '.panel-default', function (e) {
+		
+		var topDistance = 68;
+		if($('body').hasClass('admin-bar')){
+			
+			topDistance = 100;
+			
+		}
+		
+		$('html,body').animate({
+
+			scrollTop: jQuery(this).offset().top-topDistance
+			
+		}, 500); 
+	});
+	
+	// Widget - Lightbox image
+	$(".nx-fancybox").fancybox({});
+
+	// Widget - Image slider
+	$('.nx-slider').owlCarousel({
+		stagePadding: 0,
+		items: 1,
+		loop:true,
+		margin:0,
+		singleItem:true,
+		//nav:true,
+		nav:false,
+		dots:true,
+		autoplay:true,
+		autoplayTimeout:($('.nx-slider').data('interval') || 3000),
+		navText: [
+			"<i class='fa fa-angle-left'></i>",
+			"<i class='fa fa-angle-right'></i>"
+		]
+	});
 
 });