+++ core/admin/wp-bootstrap-navwalker.php
@@ -98,11 +98,11 @@ if ( ! class_exists( 'WP_Bootstrap_Navwalker' ) ) {
 				$output     .= $indent . '<li itemscope="itemscope" itemtype="https://www.schema.org/SiteNavigationElement"' . $id . $value . $class_names . '>';
 				$atts        = array();
 
-				if ( empty( $item->attr_title ) ) {
-					$atts['title'] = ! empty( $item->title ) ? strip_tags( $item->title ) : '';
-				} else {
-					$atts['title'] = $item->attr_title;
-				}
+				// if ( empty( $item->attr_title ) ) {
+					// $atts['title'] = ! empty( $item->title ) ? strip_tags( $item->title ) : '';
+				// } else {
+					// $atts['title'] = $item->attr_title;
+				// }
 
 				$atts['target'] = ! empty( $item->target ) ? $item->target : '';
 				$atts['rel']    = ! empty( $item->xfn ) ? $item->xfn : '';