+++ core/theme-support.php
@@ -8,6 +8,7 @@ add_theme_support( 'post-thumbnails' );
 add_image_size( 'fullhd', 1920, 999999 ); 
 add_image_size( 'bootstrap', 1200, 999999 );
 add_image_size( 'mobile', 768, 999999 );
+add_image_size( 'nx_thumb', 500, 312, true );
 
 // Support for HTML5
 add_theme_support( 'html5', array( 'search-form', 'comment-form', 'comment-list', 'gallery', 'caption', ) );
@@ -30,3 +31,10 @@ add_theme_support( 'custom-logo', array(
 	'flex-width'  => true,
 	'header-text' => array( 'site-title', 'site-description' ),
 ) );
+
+// Add the custom image sizes to gallery shortcode
+add_filter( 'image_size_names_choose', function ($sizes) {
+
+	return array_merge( $sizes, array('nx_thumb' => 'Thumbnail - 16:10 ratio', 'bootstrap' => 'Full size - 1200px wide', 'mobile' => 'Tablet size - 768px wide'));
+
+});